Workflows: Skip release creation if already exists. (#1416)
In case we are creating a `release-vx.y.z` branch for back-porting something to an older version, this workflow currently fails, because the commit the minor release was created from gets pushed again and so the workflow runs again, but the release already exists.

According to upstream's documentation, one can use this flag to just skip release creation in case it already exists.